requirements.txt
aws-cdk-lib>=2.0.0
aws-cdk.aws-elasticache-alpha
constructs>=10.0.0,<11.0.0
PyYAML>=6.0.0
pydantic
pytest
boto3
